% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iLaplace.R
\name{iLaplace}
\alias{iLaplace}
\title{Improved Laplace approximation for integrals of unimodal functions}
\usage{
iLaplace(fullOpt, ff, ff.gr, ff.hess,
         control = list(sp.points = 100, delta = 13, n.cores = detectCores()-1),
         extraPackages = NULL, ...)
}
\arguments{
\item{fullOpt}{A list containing the minium (to be accesed via \code{fullOpt$par}), the value of the function at the minimum (to be accessed via \code{fullOpt$objective}) and the Hessian matrix at the minimum (to be accessed via \code{fullOpt$hessian}}

\item{ff}{The minus logarithm of the integrand function (the \code{h} function, see Details).}

\item{ff.gr}{The gradient of \code{ff}, having the exact same arguments as  \code{ff}}

\item{ff.hess}{The Hessian matrix of\code{ff}, having the exact same arguments as  \code{ff}}

\item{control}{A named list of control parameters with elements \code{sp.points}, \code{delta} and \code{n.cores}. \code{sp.points} sets the number points for the spline evaluations; \code{delta} controls the length of the inteval of integration; \code{n.cores} sets the number of cores to be used for the parallel computiations. This is available for Linux/Unix systems only as it uses forking. See Details for more information.}

\item{extraPackages}{Other useful R packages to be passed to the (internally crated) clusters for the parallel computations. If required, provide package names as a vector of strings.}

\item{...}{Additional arguments to be passed to \code{ff}, \code{ff.gr} and \code{ff.hess}}
}
\value{
double, the logarithm of the integral
}
\description{
This package implements the improved Laplace approximation of Ruli et al. (2015) for multivariate integrals of user-written unimodal functions. Like the standard Laplace approximation, the improved Laplace approximation is deterministic as it replaces stochatsic integrations by numerical optimisations. However, whith respect to the Laplace approximation it involves \eqn{d}{d} (the dimensionality of the integral) scalar numerical integrations but it is more accurate, especially in those cases in which the integrand is skewed or has fat "tails". See Details below for more information. For practical examples on the use of this package please refer to the \code{iLaplaceExamples} package on \url{https://github.com/erlisR/iLaplaceExamples}.
}
\details{
\code{iLaplace} approximates integrals of the form \deqn{I = \int_{x\in\mathcal{R}^d}\exp\{-h(x)\}\,dx}{I = \int\exp(-h(x)) dx} where \eqn{-h(\cdot)}{-h()} is a concave and unimodal function, with \eqn{x}{x} being \eqn{d}{d} dimensional real vector (\eqn{d>1}{d>1}). The approximation of \eqn{I} is obtained as the ratio between the unormalised kernel \eqn{-h(x)}{-h(x)} and an approximate density function \eqn{f(x)}{f(x)}, both evaluated at the modal value \eqn{x = \hat{x}}{x = \hat{x}}. The approximate density function \eqn{f(x)}{f(x)} is obtained by resorting to the Laplace approximation for marginal densities. The normalisation of the univariate components is done as follows. First, for each of the \eqn{d}{d} dimensions, a suitable grid is fixed in the neighbourhood of \eqn{\hat x}{\hat x}. To be sure all the region with non-negligible mass is considered, the extreemes of the neighbourhood are set to \eqn{(\hat x - \delta se,\hat x + \delta se)}{(\hat x - delta*se, \hat x - delta*se)} and the grid is made of \code{sp.points} equally spaced values. Here, \eqn{se}{se} is a suitable "profile" standard error that is computed from the Hessian matrix evaluated at the modal value. \code{delta} is fixed to 13 by default. Finally, the scalar Laplace approximations for marginal densities are evaluated over the grids and the densities are reconstrcuted via the \code{\link[stats]{splinefun}} function. Lastly the splines are then normalized with the function \code{\link[stats]{integrate}} over the region constructed as above.
}
\examples{
\dontrun{

## See the examples provided in the pacakge iLaplaceExamples, which is
## an auxiliary R pacakge for iLaplace. To download it (be sure you have
## the devtools package) run from R
## devtools::install_github(erlisR/iLaplaceExamples)
## or download the source at \\url{https://github.com/erlisR/iLaplaceExamples}.

}
}
\references{
Ruli E., Sartori N. & Ventura L. (2015)
Improved Laplace approximation for marignal likelihoods.
\url{http://arxiv.org/abs/1502.06440}
}

