% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marglik.R
\name{isML}
\alias{isML}
\title{Marginal Likelihood by Importance Sampling for User-Written Functions}
\usage{
isML(logfun, nsim, theta.hat, tune, V, df, verbose)
}
\arguments{
\item{logfun}{The logarithm of the objective function}

\item{nsim}{The number of draws form the importance density}

\item{theta.hat}{The center of the proposal}

\item{tune}{A tunning value to be used to achieve the desired efficiency}

\item{V}{The scale matrix of the importance density}

\item{df}{The degrees of freedom of the importance density}

\item{verbose}{A switch which determines whether or not the progress of the sampler is printed to the screen. If verbose is greater than 0 the iteration number, and importance sampling approximation are sent to the screen every \code{verbose}th iteration}
}
\value{
double, the logarithm of the marginal likelihood
}
\description{
The function computes the marginal likelihood by importance sampling and from user-written functions.
}
\examples{
##'
}
\references{
Chib S. & Jeliazikov I. (2001).
Marginal likelihood from the Metropolis-Hastings output.
\emph{Journal of the American Statistical Association}, \bold{46},
 270--281.

Robert C. P. & Casella G. (2004).
 \emph{Monte Carlo Statistical Methods}. 2nd Edition. New York: Springer.
}
\seealso{
\code{\link[iLaplace]{iLap_an}}, \code{\link[iLaplace]{MHmcmc}},
\code{\link[iLaplace]{ChibML}}
}

