% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nLinearReg.R
\name{nLinearReg}
\alias{gradT_bod2}
\alias{gradT_lub}
\alias{grad_bod2}
\alias{grad_lub}
\alias{hessT_bod2}
\alias{hessT_lub}
\alias{hess_bod2}
\alias{hess_lub}
\alias{nLinearReg}
\alias{nlpostT_bod2}
\alias{nlpostT_lub}
\alias{nlpost_bod2}
\alias{nlpost_lub}
\title{Posterior distribution (and related quantities) for two nonlinear regression models}
\usage{
nlpost_lub(beta, lsig, y, x1, x2, n, muBeta, SigBeta, sigScale)

nlpostT_lub(beta, lsig, lnu, y, x1, x2, n, muBeta, SigBeta, sigScale)

grad_lub(beta, lsig, y, x1, x2, n, muBeta, SigBeta, sigScale)

gradT_lub(beta, lsig, lnu, y, x1, x2, n, muBeta, SigBeta, sigScale)

hess_lub(beta, lsig, y, x1, x2, n, muBeta, SigBeta, sigScale)

hessT_lub(beta, lsig, lnu, y, x1, x2, n, muBeta, SigBeta, sigScale)

nlpost_bod2(beta, lsig, y, x, n, muBeta, SigBeta, sigScale)

nlpostT_bod2(beta, lsig, lnu, y, x, n, muBeta, SigBeta, sigScale)

grad_bod2(beta, lsig, y, x, n, muBeta, SigBeta, sigScale)

gradT_bod2(beta, lsig, lnu, y, x, n, muBeta, SigBeta, sigScale)

hess_bod2(beta, lsig, y, x, n, muBeta, SigBeta, sigScale)

hessT_bod2(beta, lsig, lnu, y, x, n, muBeta, SigBeta, sigScale)
}
\arguments{
\item{beta}{The vector of regression parameters. For the Lubricant dataset beta is a 9-variate vector, for the BOD2 data beta is a bivariate vector}

\item{lsig}{The logarithm of the scale parameter}

\item{y}{The response vector}

\item{x1}{For Lubricant data only: the temperature of the lubricant}

\item{x2}{For Lubricant data only: the pressure of the lubricant}

\item{n}{The sample size}

\item{muBeta}{The center of the prior for beta. See Details}

\item{SigBeta}{The scale matrix of the prior for beta}

\item{sigScale}{For the models with Student's \eqn{t}-errors only: the value of the scale of the prior for the scale parameter.}

\item{lnu}{For the models with Student's \eqn{t}-errors only: the logarithm of the degrees of freedom}

\item{x}{For BOD2 data only: the time}
}
\value{
double

double

10-variate vector

11-variate vector

\eqn{10\times 10}{10 by 10} matrix

\eqn{11\times 11}{11 by 11} matrix

double

double

3-variate vector

4-variate vector

\eqn{3\times 3}{3 by 3} matrix

\eqn{4\times 4}{4 by 4} matrix
}
\description{
Posterior distribution, and related quantities, for two non linear regression models each based on a different dataset. For each dataset, two error distributions are considered: normal and Student's \eqn{t}. The functions ending in "lub" refer to the \code{\link[iLaplace]{Lubricant}} dataset, those with ending "bod2" refer to the \code{\link[iLaplace]{BOD2}} dataset. The functions with an inner \code{T} refer to the model based on Student's \eqn{t}-error distribution. The functions starting with \code{nlpost} give the negative log-posterior density, those starting with \code{grad} give the gradient of \code{nlpost} and the functions starting with \code{hess} provide the Hessian matrix.
}
\details{
The prior for \code{beta} is the multivariate Studnet's \eqn{t}{t} with the appropriate dimension, with location \code{muBeta}, scale matrix \code{SigBeta} and degrees of freedom equal to 5. The prior for the \code{scale} is the HalfCauchy distribution with scale \code{sigScale} and the prior for the degress of freedom is the Jeffreys' rule prior of Fonseca et al. (2008).
}
\examples{
\dontrun{

# the data
data(Lubricant)
dataObs = list(y = Lubricant$viscos, x1 = Lubricant$tempC,
               x2=Lubricant$pressure/1000)
y = Lubricant$viscos
x1 = Lubricant$tempC
x2 = Lubricant$pressure/1000
n = length(y)
muBeta = rep(0, 9)
SigBeta = matrix(0, 9,9)
diag(SigBeta) = 50
sigScale = 10
dfprop = 3

# find the maximum
init = c(10.522, 20.6, 1.464, -0.258, 0.022,
         0.394, 0.036, 46.566, -0.455, -3.195)

ff = function(x) nlpost_lub(beta = x[1:9], lsig = x[10],
                            y = y, x1 = x1, x2 = x2,
                            n = n, muBeta = muBeta,
                            SigBeta = SigBeta, sigScale = sigScale)

ff.gr = function(x) grad_lub(beta = x[1:9], lsig = x[10], y = y,
                             x1 = x1, x2 = x2, n = n, muBeta = muBeta,
                             SigBeta = SigBeta, sigScale = sigScale)

ff.hess = function(x) hess_lub(beta = x[1:9], lsig = x[10], y = y, x1 = x1,
                               x2 = x2, n = n, muBeta = muBeta,
                               SigBeta = SigBeta, sigScale = sigScale)

# find the minimum
opt.post = nlminb(init, obj=ff, gradient = ff.gr, hessian = ff.h,
                  control = list(trace = 1))
opt.post$hessian = ff.h(opt.post$par)

# draw an MCMC sample
mcmc.post = MHmcmc(logfun = function(x) -ff(x), burnin=5000,
                    mcmc=1e+5, thin=2, tune=0.9, V=solve(opt.post$hessian),
                    df=dfprop, theta.init=init, verbose=10000)
}
}
\references{
Fonseca T. C., Ferreira M. A. R. & Migon H. S. (2008)
Objective Bayesian analysis for the Student-\eqn{t}{t} regression model.
\emph{Biometrika} \bold{95}, 325--333.

Ruli E., Sartori N. & Ventura L. (2015)
Improved Laplace approximation for marignal likelihoods.
\url{http://arxiv.org/abs/1502.06440}
}
\seealso{
\code{\link[iLaplace]{iLap_an}}, \code{\link[iLaplace]{MHmcmc}}
}

