% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skewStudent.R
\name{skewStudent}
\alias{df)}
\alias{iLap_mvskt}
\alias{iLap_mvt}
\alias{iLap_mvt(m,}
\alias{skewStudent}
\title{Improved and standard Laplace approximations for multivariate skewed Student's \eqn{t}{t} densities}
\usage{
iLap_mvt(m, df)

iLap_mvskt(a, c, m, df)
}
\arguments{
\item{m}{The dimensionality}

\item{df}{The degrees of freedom}

\item{a}{The parameter a of the t/skew-t density of Jones(2002)}

\item{c}{The parameter c of the t/skew-t density of Jones(2002)}
}
\value{
a list with components \code{Lap} and \code{iLap}

a list with components \code{Lap} and \code{iLap}
}
\description{
These functions give the normalising constant of (skewed) multivariate Student's t densities approximated by the standard and the improved Laplace of Ruli et al. (2015). In particular, \code{iLap_mvt} provides the normalising constant of the \code{m}-variate Student's \eqn{t}{t} density with \code{df} degrees of freedom, centered at zero and indentiy scale matrix. \code{iLap_mvskt} provides the normalising constanto of the \code{m}-variate t/skew-t density of Jones (2002).
}
\details{
These functions use the improved and the standard Laplace approximation for computing the normalising constant of the multivariate t/skew-t density. The functions are similar to \code{\link[iLaplace]{iLap_an}}.
}
\references{
Ruli E., Sartori N. & Ventura L. (2015)
Improved Laplace approximation for marignal likelihoods.
\url{http://arxiv.org/abs/1502.06440}

Jones M. C. (2002).
Marginal replacement in multivariate densities, with application to skewing spherically symmetric distributions.
\emph{Journal of Multivariate Analysis} \bold{81}, 85--99.
}
\seealso{
\code{\link[iLaplace]{iLap_an}}
}

